requirements.txt
streamlit
python-dotenv
pdf2image
pytesseract
reportlab
requests
Pillow
